# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-nearestvalue",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — the element of `values` nearest to `value`.",
        "description": "Reads `value` and each element of `values` as a JSON integer (`i64`). If any\nis not an integer, or `values` is empty, the request is rejected with `422`.\nOtherwise the element minimizing `(element - value).abs()` is returned; ties\nresolve to the element that appears first.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearestValueResponse"
                }
              }
            }
          },
          "422": {
            "description": "value or an element is not an integer, or values is empty"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value",
          "values"
        ],
        "properties": {
          "value": {
            "type": "object",
            "description": "The reference value. Must be a JSON integer (i64)."
          },
          "values": {
            "type": "object",
            "description": "The non-empty list of candidates. Every element must be a JSON integer."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "NearestValueResponse": {
        "type": "object",
        "required": [
          "value",
          "values",
          "result"
        ],
        "properties": {
          "result": {
            "type": "integer",
            "format": "int64"
          },
          "value": {
            "type": "integer",
            "format": "int64"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      }
    }
  }
}